TK-33585: Update Artifact Manager Service Context to remove the deployment directory on shutdown.  Also some other minor sonar refactorings in this checkin.
