+++ project-set/components/client-auth/src/main/java/com/rackspace/papi/components/clientauth/ClientAuthenticationHandlerFactory.java
@@ -1,5 +1,6 @@
 package com.rackspace.papi.components.clientauth;
 
+import com.rackspace.papi.components.clientauth.common.AuthModule;
 import com.rackspace.papi.components.clientauth.common.UriMatcher;
 import com.rackspace.papi.commons.config.manager.UpdateListener;
