+++ project-set/components/client-auth/src/main/java/com/rackspace/papi/components/clientauth/openstack/v1_0/OpenStackAuthenticationHandler.java
@@ -5,7 +5,7 @@ import com.rackspace.papi.commons.util.regex.KeyedRegexExtractor;
 import com.rackspace.auth.openstack.ids.CachableUserInfo;
 import com.rackspace.auth.openstack.ids.OpenStackAuthenticationService;
 import com.rackspace.docs.identity.api.ext.rax_ksgrp.v1.Groups;
-import com.rackspace.papi.components.clientauth.AuthModule;
+import com.rackspace.papi.components.clientauth.common.AuthModule;
 import com.rackspace.papi.components.clientauth.common.UriMatcher;
 import com.rackspace.papi.filter.logic.common.AbstractFilterLogicHandler;
