+++ project-set/components/client-auth/src/main/java/com/rackspace/papi/components/clientauth/rackspace/v1_1/RackspaceAuthenticationHandlerFactory.java
@@ -2,7 +2,7 @@ package com.rackspace.papi.components.clientauth.rackspace.v1_1;
 
 import com.rackspace.auth.v1_1.AuthenticationServiceClient;
 import com.rackspace.auth.v1_1.AuthenticationServiceClientFactory;
-import com.rackspace.papi.components.clientauth.AuthModule;
+import com.rackspace.papi.components.clientauth.common.AuthModule;
 import com.rackspace.papi.commons.util.regex.KeyedRegexExtractor;
 import com.rackspace.papi.components.clientauth.common.UriMatcher;
 import com.rackspace.papi.components.clientauth.config.ClientAuthConfig;
