+++ project-set/core/core-lib/src/main/java/com/rackspace/papi/service/deploy/ArtifactManagerServiceContext.java
@@ -9,14 +9,20 @@ import com.rackspace.papi.service.event.PowerFilterEvent;
 import com.rackspace.papi.service.event.common.Event;
 import com.rackspace.papi.service.event.common.EventService;
 import com.rackspace.papi.service.event.listener.SingleFireEventListener;
+import org.slf4j.Logger;
+
 import javax.servlet.ServletContext;
 import javax.servlet.ServletContextEvent;
+import java.io.File;
 
 public class ArtifactManagerServiceContext implements ServiceContext<ArtifactManager> {
+   private static final Logger LOG = org.slf4j.LoggerFactory.getLogger(ArtifactManagerServiceContext.class);
 
    public static final String SERVICE_NAME = "powerapi:/kernel/artifact-deployment";
    private DestroyableThreadWrapper watcherThread;
    private ArtifactManager artifactManager;
+
+   private ContainerConfigurationListener containerCfgListener = null;
 
    @Override
    public String getServiceName() {
@@ -35,7 +41,7 @@ public class ArtifactManagerServiceContext implements ServiceContext<ArtifactMan
 
       final EventService eventManagerReference = contextAdapter.eventService();
 
-      final ContainerConfigurationListener containerCfgListener = new ContainerConfigurationListener(eventManagerReference);
+      containerCfgListener = new ContainerConfigurationListener(eventManagerReference);
       watcherThread = new DestroyableThreadWrapper(contextAdapter.threadingService().newThread(containerCfgListener.getDirWatcher(), "Artifact Watcher Thread"), containerCfgListener.getDirWatcher());
 
       contextAdapter.configurationService().subscribeTo("container.cfg.xml", containerCfgListener, ContainerConfiguration.class);
@@ -58,8 +64,23 @@ public class ArtifactManagerServiceContext implements ServiceContext<ArtifactMan
       try {
          final EventService eventManagerReference = ServletContextHelper.getInstance().getPowerApiContext(sce.getServletContext()).eventService();
          eventManagerReference.squelch(artifactManager, ApplicationArtifactEvent.class);
+
+         if (containerCfgListener.isAutoClean()) {
+            delete(containerCfgListener.getUnpacker().getDeploymentDirectory());
+         }
       } finally {
          watcherThread.destroy();
+      }
+   }
+
+   private void delete(File f) {
+      if (f.isDirectory()) {
+         for (File c : f.listFiles())
+            delete(c);
+      }
+
+      if(!f.delete()) {
+         LOG.warn("Failure to delete file " + f.getName() + " on repose shutdown.");
       }
    }
 }
